+++ BE/musicspot/src/journey/schema/journey.schema.ts
