+++ BE/musicspot/src/spot/schema/spot.schema.ts
