+++ BE/musicspot/src/user/serivce/user.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UserService } from './user.service';
 import { getModelToken } from '@nestjs/mongoose';
-import { User, UserSchema } from './user.schema';
+import { User, UserSchema } from '../schema/user.schema';
 import mongoose from 'mongoose';
 
 describe('UserService', () => {
